move related types to ReadHandler module
