+++ dnp3/src/master/handler.rs
@@ -1,7 +1,6 @@
 use std::net::SocketAddr;
 use std::time::{Duration, SystemTime};
 
-use crate::app::attr::*;
 use crate::app::*;
 
 use crate::decode::DecodeLevel;
@@ -594,69 +593,4 @@ pub trait AssociationInformation: Send + Sync {
 
     /// Called when an unsolicited response is received
     fn unsolicited_response(&mut self, _is_duplicate: bool, _seq: Sequence) {}
-}
-
-/// Information about the object header and specific variation
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub struct HeaderInfo {
-    /// underlying variation in the response
-    pub variation: Variation,
-    /// qualifier code used in the response
-    pub qualifier: QualifierCode,
-    /// true if the received variation is an event type, false otherwise
-    pub is_event: bool,
-    /// true if a flags byte is present on the underlying variation, false otherwise
-    pub has_flags: bool,
-}
-
-impl HeaderInfo {
-    pub(crate) fn new(
-        variation: Variation,
-        qualifier: QualifierCode,
-        is_event: bool,
-        has_flags: bool,
-    ) -> Self {
-        Self {
-            variation,
-            qualifier,
-            is_event,
-            has_flags,
-        }
-    }
-}
-
-/// Describes the source of a read event
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
-pub enum ReadType {
-    /// Startup integrity poll
-    StartupIntegrity,
-    /// Unsolicited message
-    Unsolicited,
-    /// Single poll requested by the user
-    SinglePoll,
-    /// Periodic poll configured by the user
-    PeriodicPoll,
-}
-
-pub(crate) fn handle_attribute(
-    var: Variation,
-    qualifier: QualifierCode,
-    attr: &Option<Attribute>,
-    handler: &mut dyn ReadHandler,
-) {
-    if let Some(attr) = attr {
-        match AnyAttribute::try_from(attr) {
-            Ok(attr) => {
-                handler
-                    .handle_device_attribute(HeaderInfo::new(var, qualifier, false, false), attr);
-            }
-            Err(err) => {
-                tracing::warn!(
-                    "Expected attribute type {:?} but received {:?}",
-                    err.expected,
-                    err.actual
-                );
-            }
-        }
-    }
 }
